requirements.txt
pip==24.0
reflex==0.4.3
python-dotenv==1.0.1
supabase==2.3.5
configcat-client==9.0.2
pytz==2024.1
